+++ src/timber_design/ghpython/components/CT_Bake_BeamBoxMap/code.py
@@ -1,4 +1,3 @@
-# r: timber_design>=0.1.0
 # flake8: noqa
 import math
 import random